requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.6.0
attrs==23.2.0
autopep8==2.1.0
bcrypt==4.1.2
black==24.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.4
decorator==5.1.1
Deprecated==1.2.14
environs==11.0.0
exceptiongroup==1.2.0
execnet==2.0.2
fabric==3.2.2
fake==0.8
Faker==23.1.0
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
invoke==2.2.0
marshmallow==3.21.1
mypy-extensions==1.0.0
packaging==23.2
paramiko==3.4.0
pathspec==0.12.1
platformdirs==4.2.0
playwright==1.41.2
pluggy==1.4.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyee==11.0.1
PyNaCl==1.5.0
pytest==8.0.0
pytest-assume==2.4.3
pytest-repeat==0.9.3
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rootpath==0.1.0
setuptools==69.2.0
six==1.16.0
swagger-coverage==3.4.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.0
wrapt==1.16.0
